Replace cfree() with free()
cfree was calloc's analogue to free. But at least in Unix, it's always
just been an alias for free. I'm not an expert on early C, so maybe some
architecture actually depended on separate free behavior for calloc'd
blocks? It appeared in Bell Labs Unix in V6 and was just a wrapper around
free(), so it's not clear if this was for compatibility with another
system or just for the sake of symmetry.

The function lived on at least into some versions of SunOS, and from
there it was implemented in glibc for compatibility with SunOS. It never
made it into ANSI/ISO C or POSIX, but it did end up being standardized
in iBCS somehow. No one cares about iBCS anymore and glibc dropped cfree
in 2017.