Rename index to sub1_index
index() was added in Version 6 Unix and does what strchr does now, which
is not what this function does (it returns an integer index into the
string, not a pointer). I'm not super clear how the linker worked on a
PDP-11 so I'll be honest that this one is a little confusing as far as
how it ever worked.