Remove a useless allocation
Besides the int vs. pointer incompatibility, the only purpose of this
block seems to be to check whether there is enough memory available.
It's allocated, checked for not NULL, then freed, and the actual
allocations happen a couple lines down. Maybe this was an artifact of
some sort of home-grown slab allocator. I don't know, but it's in the
way now and doesn't actually do anything, ditch it.