Lightly modernize the yacc file
Remove two syntax forms that are described in "Yacc: Yet Another
Compiler-Compiler" as "Old Features Supported but not Encouraged": the
%{ ... }% code block at the beginning of the rules section, and the
`RULE = { ... }` form of defining an action.

For the code block in rules, the content was all global variables that
aren't actually used as globals. Add variable definitions to the start
of the appropriate action blocks and remove the extra code block.

For the `RULE = { ... }` form, remove the equal signs.