Add the ratfor driver from BSD
This isn't included in any of the Bell Labs Unix distributions on The
Unix Heritage Project. I'm not sure if this is because of incomplete
tapes or because no one at AT&T cared about ratfor. I don't care about
ratfor either, I just think it's weird that it's missing. This copy is
from 2.11 BSD.